pyproject.toml
[tool.poetry]
name = "myshop"
version = "0.1.0"
description = ""
authors = ["Artem <avtorca4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
pillow = "^9.4.0"
django-bootstrap4 = "^22.3"
yookassa = "^2.3.5"
python-dotenv = "^0.21.1"
django-celery-beat = "^2.4.0"
celery = "^5.2.7"
redis = "^4.5.1"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
